requirements.txt
Flask==2.3.2
requests>=2.28.2
beautifulsoup4>=4.13.4
nltk>=3.7
googlesearch-python>=1.1.0
transformers>=4.28.1
flask-caching>=1.11.0
wikipedia-api>=0.5.4
Werkzeug==2.3.1
torch>=2.0.1
flax>=0.6.0
wikipedia>=1.4.0
aiohttp>=3.8.4
